statUI
스텟올리기

+++ Assets/02.Script/Player/PlayerStatus.cs
@@ -15,8 +15,8 @@ public class PlayerStatus : MonoBehaviour, ICharacterStatus
     [SerializeField] private int _accuracy;
     [SerializeField] private int _evasion;
 
-    [SerializeField] private float _goldAscend;
-    [SerializeField] private float _expAscend;
+    [SerializeField] private int _goldAscend;
+    [SerializeField] private int _expAscend;
 
     public int MaxHp { get => _maxHp; set => _maxHp = value; }
     public int Power { get => _power; set => _power = value; }
@@ -28,6 +28,6 @@ public class PlayerStatus : MonoBehaviour, ICharacterStatus
     public int Accuracy { get => _accuracy; set => _accuracy = value; }
     public int Evasion { get => _evasion; set => _evasion = value; }
 
-    public float GoldAscend => _goldAscend;
-    public float ExpAscend => _expAscend;
+    public int GoldAscend => _goldAscend;
+    public int ExpAscend => _expAscend;
 }
